Moves RegexReplacer into StringFunctions
